Allow Spark/Hadoop version to be set during build.
I have modified the packaging so that Spark and Hadoop versions are
parametrized. With this, the versions can be set on the Maven command line
during build, e.g.:

mvn package -Dhadoop.version=2.6.2 -Dspark.version=1.5.2

This resolves an issue we were having in the toil/ADAM pipeline, since the
version of HDFS we were using did not match the version of HDFS that Conductor
was built against.